move statereader to storage-api
